fix(library-solidity): sync bridge wrappers with updated host send/quote ABI
host #2880 dropped the options parameter from ConfidentialBridge.send/quote and narrowed lzComposeGas to uint64. Update the IConfidentialBridge mirror, the FHE.bridge/quoteBridge overloads (incl. the low-level bytes32 escape hatch), Impl, and the ConfidentialOAppSender helper accordingly, plus the example/harness/mocks/tests. The bridge now always builds its own LayerZero options and enforces a per-dstEid minimum compose gas.

+++ library-solidity/examples/bridge/ConfidentialOFTViaLib.sol
@@ -60,7 +60,7 @@ contract ConfidentialOFTViaLib is Ownable2Step, ConfidentialOAppSender, Confiden
         uint32 dstEid,
         euint64 amount,
         address recipient,
-        uint128 mintComposeGas
+        uint64 mintComposeGas
     ) external payable returns (MessagingReceipt memory) {
         if (!FHE.isSenderAllowed(amount)) revert UnauthorizedUseOfEncryptedAmount(amount, msg.sender);
 
@@ -81,7 +81,7 @@ contract ConfidentialOFTViaLib is Ownable2Step, ConfidentialOAppSender, Confiden
         uint32 dstEid,
         euint64 amount,
         address recipient,
-        uint128 mintComposeGas
+        uint64 mintComposeGas
     ) external view returns (MessagingFee memory) {
         return _quoteBridge(dstEid, abi.encode(recipient), amount, mintComposeGas);
     }
